Changed a lot

+++ TabletApp/TabletApp/Models/MenuItem.cs
@@ -17,7 +17,7 @@ namespace TabletApp.Models
         public string name { get; set; }
         public string picture { get; set; }
         public string description { get; set; }
-        public float price { get; set; }
+        public double price { get; set; }
         public string nutrition { get; set; }
         public string item_type { get; set; }
         public string category { get; set; }
